move fileops, flesh out, reorder methods
